backend/pyproject.toml
[project]
name = "mramor-and-granit"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

requires-python = ">=3.12,<3.15"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-extra-types (>=2.11.0,<3.0.0)",
    "phonenumbers (>=9.0.22,<10.0.0)",
    "aiogram (>=3.24.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
package-mode = false
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
